szrpanalitikchatvk/requirements.txt
# Основные зависимости для упрощенного VK бота
aiohttp>=3.8.0
aiogram>=3.1.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
loguru>=0.7.0
